Demostrate the user of Vaadin Components
Datasource: options: BeanItemContainer, which takes a list of
collection.

This is an Vaadin Application that demostrate Grid and ComboBox that
bind with the same datasource
When Click the Item in the Grid and click delete button, it will remove
the item from the datasource
and update the content of Grid and ComboBox automatically.

The TextField component that takes the search input from the user and
modify the datasource accordingly. When the reset button is pressed, the
datasource is reset to it's orginally state.

The list of collection did not modified onces the datasource item is
removed.